pyproject.toml
[tool.poetry]
name = "llm-solution-generator"
version = "0.1.0"
description = "A Flask app that generates and runs solutions using LLM"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 